+++ packages/engineer/src/application-proxy-service.ts
@@ -7,13 +7,13 @@ import {
     ICompilerOptions,
     OverrideConfig,
     INpmPackage,
+    IFeatureDefinition,
 } from '@wixc3/engine-scripts';
 import type { SetMultiMap } from '@wixc3/engine-core';
 import performance from '@wixc3/cross-performance';
 import type {
     IConfigDefinition,
     IExternalDefinition,
-    IFeatureDefinition,
     LaunchEnvironmentMode,
     NodeEnvironmentsManager,
 } from '@wixc3/engine-runtime-node';
