+++ packages/engineer/src/feature/gui.feature.ts
@@ -1,5 +1,6 @@
 import { Feature, Environment, COM, Config } from '@wixc3/engine-core';
-import type { IExtenalFeatureDescriptor, IFeatureDefinition } from '@wixc3/engine-runtime-node';
+import type { IExtenalFeatureDescriptor } from '@wixc3/engine-runtime-node';
+import type { IFeatureDefinition } from '@wixc3/engine-scripts';
 import buildFeature from './dev-server.feature';
 
 export const mainDashboardEnv = new Environment('main-dashboard', 'window', 'single');
