+++ packages/scripts/src/create-entrypoint.ts
@@ -1,7 +1,8 @@
 import type { SetMultiMap, TopLevelConfig } from '@wixc3/engine-core';
-import type { IFeatureDefinition, IConfigDefinition } from '@wixc3/engine-runtime-node';
+import type { IConfigDefinition } from '@wixc3/engine-runtime-node';
 import { parse } from 'path';
 import { CONFIG_QUERY_PARAM, FEATURE_QUERY_PARAM } from './build-constants';
+import type { IFeatureDefinition } from './types';
 
 const { stringify } = JSON;
 const topLevelConfigLoaderPath = require.resolve('./top-level-config-loader');
@@ -69,6 +70,7 @@ export interface LoadStatementArguments
 //#region entry points
 
 export function createExternalBrowserEntrypoint(args: WebpackFeatureLoaderArguments) {
+
     return `
     import { getTopWindow } from ${JSON.stringify(require.resolve('@wixc3/engine-core'))};
     const topWindow = getTopWindow(typeof self !== 'undefined' ? self : window);
@@ -85,7 +87,7 @@ export function createExternalNodeEntrypoint(args: ExternalEntrypoint) {
         ...args,
         target: 'node',
         loadStatement: nodeImportStatement,
-    })}
+    })} 
 }
     `;
 }
@@ -127,9 +129,8 @@ async function main() {
     const topWindow = getTopWindow(currentWindow);
     const options = new URLSearchParams(topWindow.location.search);
 
-    const publicPath = options.has('publicPath') ? options.get('publicPath') : ${
-        typeof publicPath === 'string' ? JSON.stringify(publicPath) : '__webpack_public_path__'
-    };
+    const publicPath = options.has('publicPath') ? options.get('publicPath') : ${typeof publicPath === 'string' ? JSON.stringify(publicPath) : '__webpack_public_path__'
+        };
     __webpack_public_path__= publicPath;
 
     const featureName = options.get('${FEATURE_QUERY_PARAM}') || ${stringify(featureName)};
@@ -166,9 +167,8 @@ main().catch(console.error);
 
 //#region webpack import statements
 export function webpackImportStatement({ moduleIdentifier, filePath, eagerEntrypoint }: LoadStatementArguments) {
-    return `await import(/* webpackChunkName: "${moduleIdentifier}" */${
-        eagerEntrypoint ? ` /* webpackMode: 'eager' */` : ''
-    } ${stringify(filePath)});`;
+    return `await import(/* webpackChunkName: "${moduleIdentifier}" */${eagerEntrypoint ? ` /* webpackMode: 'eager' */` : ''
+        } ${stringify(filePath)});`;
 }
 
 export function nodeImportStatement({ filePath }: LoadStatementArguments) {
@@ -221,12 +221,12 @@ function loadEnvAndContextFiles({
             usesResolvedContexts = true;
             loadStatements.push(`if (resolvedContexts[${JSON.stringify(envName)}] === ${JSON.stringify(childEnvName)}) {
                 ${loadStatement({
-                    moduleIdentifier: name,
-                    filePath: contextFilePath,
-                    directoryPath,
-                    packageName,
-                    eagerEntrypoint,
-                })};
+                moduleIdentifier: name,
+                filePath: contextFilePath,
+                directoryPath,
+                packageName,
+                eagerEntrypoint,
+            })};
             }`);
         }
         const preloadFilePath = preloadFilePaths?.[`${envName}/${childEnvName}`];
@@ -235,12 +235,12 @@ function loadEnvAndContextFiles({
             usesResolvedContexts = true;
             preloadStatements.push(`if (resolvedContexts[${stringify(envName)}] === ${stringify(childEnvName)}) {
                 ${webpackImportStatement({
-                    directoryPath,
-                    filePath: preloadFilePath,
-                    moduleIdentifier: name,
-                    packageName,
-                    eagerEntrypoint,
-                })};
+                directoryPath,
+                filePath: preloadFilePath,
+                moduleIdentifier: name,
+                packageName,
+                eagerEntrypoint,
+            })};
             }`);
         }
     }
@@ -288,17 +288,16 @@ function createLoaderInterface(args: WebpackFeatureLoaderArguments) {
                 async load(${usesResolvedContexts ? 'resolvedContexts' : ''}) {
                     ${loadStatements.length ? loadStatements.join(';\n') : ''}
                     const featureModule = ${loadStatement({
-                        moduleIdentifier: `[feature]${name}`,
-                        filePath,
-                        directoryPath,
-                        packageName,
-                        eagerEntrypoint,
-                    })};
-                    ${
-                        target !== 'node'
-                            ? `self.${createExternalFeatureMapping(packageName, filePath)} = featureModule;`
-                            : ''
-                    }
+        moduleIdentifier: `[feature]${name}`,
+        filePath,
+        directoryPath,
+        packageName,
+        eagerEntrypoint,
+    })};
+                    ${target !== 'node'
+            ? `self.${createExternalFeatureMapping(packageName, filePath)} = featureModule;`
+            : ''
+        }
                     return featureModule.default;
                 },
                 async preload(${usesResolvedContexts ? 'resolvedContexts' : ''}) {
@@ -355,11 +354,10 @@ function createConfigLoaders(configs: Record<string, IConfigFileMapping[]>) {
 }
 
 function loadConfigFile(filePath: string, scopedName: string, configEnvName: string | undefined): string {
-    return `import(/* webpackChunkName: "[config]${scopedName}${
-        configEnvName ?? ''
-    }" */ /* webpackMode: 'eager' */ ${JSON.stringify(
-        topLevelConfigLoaderPath + `?scopedName=${scopedName}&envName=${configEnvName!}!` + filePath
-    )})`;
+    return `import(/* webpackChunkName: "[config]${scopedName}${configEnvName ?? ''
+        }" */ /* webpackMode: 'eager' */ ${JSON.stringify(
+            topLevelConfigLoaderPath + `?scopedName=${scopedName}&envName=${configEnvName!}!` + filePath
+        )})`;
 }
 //#endregion
 
@@ -395,11 +393,10 @@ function loadExternalFeatures(target: 'web' | 'webworker' | 'electron-renderer',
     
     ${addExternalsEventListenerForParentEnvironments(externalsFilePath)}
     
-    const fetchedExternalFeatures = ${
-        target === 'electron-renderer'
+    const fetchedExternalFeatures = ${target === 'electron-renderer'
             ? fetchFeaturesFromElectronProcess(externalsFilePath)
             : fetchExternalFeaturesInBrowser(externalsFilePath)
-    };
+        };
     externalFeatures.push(...fetchedExternalFeatures)
     
     if(externalFeatures.length) {
