+++ packages/scripts/src/create-webpack-configs.ts
@@ -13,11 +13,11 @@ import {
 import { WebpackScriptAttributesPlugin } from './webpack-html-attributes-plugins';
 import { createVirtualEntries } from './virtual-modules-loader';
 import type {
-    getResolvedEnvironments,
     IConfigDefinition,
-    IFeatureDefinition,
     TopLevelConfigProvider,
 } from '@wixc3/engine-runtime-node';
+import type { getResolvedEnvironments } from './utils/environments';
+import type { IFeatureDefinition } from './types';
 
 export interface ICreateWebpackConfigsOptions {
     baseConfig?: webpack.Configuration;
@@ -291,22 +291,22 @@ export function createWebpackConfigForExternalFeature({
 
 const extractExternals =
     (featurePath: string, ignoredRequests: string[]) =>
-    ({ context, request }: { context?: string; request?: string }, cb: (e?: Error, target?: string) => void) => {
-        try {
-            if (!request || !context || ignoredRequests.includes(request) || request.includes('!=!')) {
-                return cb();
-            }
-            const resolvedRequest = require.resolve(request, { paths: [context] });
-            if (resolvedRequest !== featurePath && fs.basename(resolvedRequest).includes('.feature.')) {
-                const packageJson = fs.findClosestFileSync(fs.dirname(resolvedRequest), 'package.json');
-                if (!packageJson) {
-                    throw new Error(`could not find package.json for ${resolvedRequest}`);
+        ({ context, request }: { context?: string; request?: string }, cb: (e?: Error, target?: string) => void) => {
+            try {
+                if (!request || !context || ignoredRequests.includes(request) || request.includes('!=!')) {
+                    return cb();
                 }
-                const { name } = fs.readJsonFileSync(packageJson) as { name: string };
-                return cb(undefined, createExternalFeatureMapping(name, resolvedRequest));
+                const resolvedRequest = require.resolve(request, { paths: [context] });
+                if (resolvedRequest !== featurePath && fs.basename(resolvedRequest).includes('.feature.')) {
+                    const packageJson = fs.findClosestFileSync(fs.dirname(resolvedRequest), 'package.json');
+                    if (!packageJson) {
+                        throw new Error(`could not find package.json for ${resolvedRequest}`);
+                    }
+                    const { name } = fs.readJsonFileSync(packageJson) as { name: string };
+                    return cb(undefined, createExternalFeatureMapping(name, resolvedRequest));
+                }
+                cb();
+            } catch (err) {
+                cb(err);
             }
-            cb();
-        } catch (err) {
-            cb(err);
-        }
-    };
+        };
