+++ packages/scripts/src/utils/index.ts
@@ -4,3 +4,4 @@ export * from './resolve-packages';
 export * from './string-utils';
 export * from './external-features';
 export * from './paths';
+export * from './environments';
